+++ fyl_server/.idea/compiler.xml
@@ -8,7 +8,6 @@
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.30/f195ee86e6c896ea47a1d39defbe20eb59cd149d/lombok-1.18.30.jar" />
         </processorPath>
-        <module name="fyl_server.main" />
         <module name="fyl_server.test" />
       </profile>
     </annotationProcessing>
