+++ fyl_server/src/main/resources/application.properties
@@ -11,7 +11,6 @@ jdbc.password=0805
 
 
 
-
 # JPA ??
 spring.jpa.hibernate.ddl-auto=update  # 'create', 'create-drop', 'update', 'validate' ? ??
 spring.jpa.show-sql=true             # SQL ??? ??? ??
